fix(ci): python-coverage-shards — align fail-closed tests with os.open/ls-files and split oversized residue modules
Residue probes now read via os.open and list nested untracked with ls-files -o; update tests to exercise those seams, and extract IO helpers / part_004 so files stay under the 1500-line gate.
